requirements.txt
flask
pandas
werkzeug
gunicorn
openai
python-dotenv
markdown
couchbase
